feat: add admin guard

+++ backEnd/app/Services/AuthService.php
@@ -8,15 +8,22 @@ use Illuminate\Validation\ValidationException;
 
 class AuthService
 {
+    protected $model;
+    protected $guard;
+    public function __construct($model,$guard = 'web')
+    {
+        $this->model = $model;
+        $this->guard = $guard;
+    }
     public function register(array $data)
     {
-        $user = User::create([
+        $user = $this->model::create([
             'name' => $data['name'],
             'email' => $data['email'],
             'password' => bcrypt($data['password']),
         ]);
 
-        $token = $user->createToken('auth-token')->plainTextToken;
+        $token = $user->createToken($this->guard.'-token')->plainTextToken;
 
         return [
             'user' => $user,
@@ -26,13 +33,13 @@ class AuthService
 
     public function login(array $data)
     {
-        if (!Auth::attempt($data)) {
+        if (!Auth::guard($this->guard)->attempt($data)) {
             return null;
         }
 
-        $user = User::where('email', $data['email'])->first();
+        $user = $this->model::where('email', $data['email'])->first();
 
-        $token = $user->createToken('auth-token')->plainTextToken;
+        $token = $user->createToken($this->guard.'-token')->plainTextToken;
 
         return [
             'user' => $user,
@@ -40,8 +47,12 @@ class AuthService
         ];
     }
 
-    public function logout($user, bool $allDevices = false)
+    public function logout(bool $allDevices = false)
     {
+        $user = Auth::guard($this->guard)->user();
+        if (!$user) {
+            return;
+        }
         if ($allDevices) {
             $user->tokens()->delete();
         } else {
